pyproject.toml
[project]
name = 'jpl.edrn.pubnot'
dynamic = ['version']
requires-python = '>=3.9'
dependencies = [
    'rdflib ~= 7.0.0',
    'biopython ~= 1.84'
]
authors = [
    {name = 'Sean Kelly', email='kelly@seankelly.biz'}
]
description = 'Publication change notifier for the Early Detection Research Network'
readme = 'README.md'
keywords = ['edrn', 'cancer', 'biomarker', 'publication', 'pubmed', 'entrez', 'maureen']
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Healthcare Industry',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: Apache Software License',
    'Natural Language :: English',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3',
]
license = {file = 'LICENSE.md'}


[project.urls]
Homepage = 'https://github.com/EDRN/pubnot'
Issues = 'https://github.com/EDRN/pubnot/issues'


[project.scripts]
edrn-pubnot = 'jpl.edrn.pubnot.main:main'


[tool.hatch.version]
path = 'src/jpl/edrn/pubnot/VERSION.txt'
pattern = '(?P<version>.+)'


[tool.hatch.build.targets.wheel]
packages = ['src/jpl']


[tool.hatch.metadata]
allow-direct-references = true


[build-system]
requires = ['hatchling']
build-backend = 'hatchling.build'
